# doc/api.yml
openapi: 3.0.3
info:
  title: Async Email API
  version: 1.0.0
  description: Course subscription API

paths:
  /auth/signup:
    post:
      summary: Register a new user account
      operationId: signUp
      tags: [Auth]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SignUpRequest'
      responses:
        '201':
          description: Account created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorBody'
        '409':
          description: Username or email already taken
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorBody'

  /auth/login:
    post:
      summary: Authenticate and receive a JWT token
      operationId: logIn
      tags: [Auth]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: Authentication successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '401':
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorBody'

  /users/{userId}/courses/{courseId}:
    post:
      summary: Subscribe a user to a course
      operationId: subscribeToCourse
      tags: [UserCourses]
      security:
        - BearerAuth: []
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: courseId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '201':
          description: Subscription created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserCourseResponse'
        '404':
          description: User or course not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorBody'
        '409':
          description: User already subscribed to this course
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorBody'

  /data:
    get:
      summary: List all submitted image data with pagination
      operationId: listImageData
      tags: [Data]
      parameters:
        - name: email
          in: query
          required: false
          schema:
            type: string
            format: email
          description: Filter by email (exact match)
        - name: filename
          in: query
          required: false
          schema:
            type: string
          description: Filter by filename (partial match)
        - name: page
          in: query
          required: false
          schema:
            type: integer
            minimum: 1
            default: 1
          description: Page number (1-indexed)
        - name: size
          in: query
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 10
          description: Number of images per page
      responses:
        '200':
          description: Paginated list of image records
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ImageListResponse'
        '400':
          description: Bad request — invalid query parameters
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorBody'
        '422':
          description: Unprocessable entity — validation error on filter values
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorBody'

    post:
      summary: Submit an image for processing (JPEG or PNG)
      operationId: submitImageData
      tags: [Data]
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [file, email]
              properties:
                file:
                  type: string
                  format: binary
                  description: Image file (JPEG or PNG)
                email:
                  type: string
                  format: email
                  description: Email to receive the processed image link
      responses:
        '201':
          description: Image submitted and record created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ImageResponse'
        '400':
          description: Invalid input — unsupported format or missing fields
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorBody'
        '422':
          description: Unprocessable entity — validation error on submitted data
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorBody'

  /courses:
    get:
      summary: List all available courses with optional filters
      operationId: listCourses
      tags: [Courses]
      parameters:
        - name: title
          in: query
          required: false
          schema:
            type: string
          description: Filter by course title (partial match)
        - name: startDate
          in: query
          required: false
          schema:
            type: string
            format: date-time
          description: Filter by minimum start date (courses starting on or after)
        - name: endDate
          in: query
          required: false
          schema:
            type: string
            format: date-time
          description: Filter by maximum end date (courses ending on or before)
        - name: price
          in: query
          required: false
          schema:
            type: number
            format: double
          description: Maximum price the user can afford (filters courses with price ≤ this value)
        - name: page
          in: query
          required: false
          schema:
            type: integer
            minimum: 1
            default: 1
          description: Page number (1-indexed)
        - name: pageSize
          in: query
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 10
          description: Number of courses per page
      responses:
        '200':
          description: List of courses
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CourseListResponse'
        '400':
          description: Bad request — invalid query parameters
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorBody'
        '422':
          description: Unprocessable entity — validation error on filter values
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorBody'

components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    SignUpRequest:
      type: object
      required: [firstName, lastName, username, email, password, confirmPassword]
      properties:
        firstName:
          type: string
          maxLength: 100
        lastName:
          type: string
          maxLength: 100
        username:
          type: string
          maxLength: 50
        email:
          type: string
          format: email
          maxLength: 100
        password:
          type: string
          minLength: 8
          maxLength: 128
        confirmPassword:
          type: string
          minLength: 8
          maxLength: 128

    LoginRequest:
      type: object
      required: [username, password]
      properties:
        username:
          type: string
        password:
          type: string

    UserResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        firstName:
          type: string
        lastName:
          type: string
        username:
          type: string
        email:
          type: string
          format: email
        role:
          $ref: '#/components/schemas/UserRole'

    AuthResponse:
      type: object
      properties:
        token:
          type: string
          description: JWT Bearer token
        user:
          $ref: '#/components/schemas/UserResponse'

    UserRole:
      type: string
      enum: [ADMIN, CUSTOMER]

    UserCourseResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        userId:
          type: string
          format: uuid
        courseId:
          type: string
          format: uuid
        subscribedAt:
          type: string
          format: date-time

    CourseListResponse:
      type: object
      required: [data, meta]
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/CourseResponse'
        meta:
          $ref: '#/components/schemas/Meta'

    Meta:
      type: object
      properties:
        page:
          type: integer
          description: Current page number
        size:
          type: integer
          description: Number of items per page
        total:
          type: integer
          description: Total number of courses matching the filters

    CourseResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        title:
          type: string
        startDate:
          type: string
          format: date-time
        endDate:
          type: string
          format: date-time
        price:
          type: number
          format: double

    ImageResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        filename:
          type: string
        email:
          type: string
          format: email
        createdAt:
          type: string
          format: date-time

    ImageListResponse:
      type: object
      required: [data, meta]
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/ImageResponse'
        meta:
          $ref: '#/components/schemas/Meta'

    ErrorBody:
      type: object
      properties:
        status:
          type: integer
        error:
          type: string
        message:
          type: string
        timestamp:
          type: string
          format: date-time